refactor: abstract pixel handling
